+++ src/frontend/app/api/fetch-token/route.ts
@@ -3,9 +3,9 @@ import { set_token } from '@/lib/tokencache'
 import { timestamp_string } from '@/lib/utils'
 
 export async function POST() {
-  const token_url: string = process.env.API_TOKEN_URL as string;
-  const username: string = process.env.API_USER as string;
-  const password: string = process.env.API_PASS as string;
+  const token_url: string = process.env.API_TOKEN_URL as string
+  const username: string = process.env.API_USER as string
+  const password: string = process.env.API_PASS as string
 
   const form_data = new URLSearchParams({
     'grant_type': 'password',
@@ -36,6 +36,6 @@ export async function POST() {
     NextResponse.json({ error: 'Fetch Token Response Rejected!' }, { status: 500 })
   })
   let token = await tokendata.access_token
-  set_token(token);
+  set_token(token)
   return NextResponse.json(tokendata, { status: 200 })
 }