+++ src/frontend/components/AppMenu.tsx
@@ -1,55 +1,71 @@
-'use client';
+'use client'
 
-import { BookUser, FileLock, HomeIcon, QrCode, ReceiptText, ScrollText, Settings } from "lucide-react";
-import { Command, CommandGroup, CommandItem, CommandList } from "./ui/command";
-import Link from "next/link";
+import { HomeIcon, QrCode, Package, Puzzle, BookMarked, Asterisk, BookText } from 'lucide-react'
+import { Command, CommandGroup, CommandItem, CommandList } from './ui/command'
+import Link from 'next/link'
 
 export default function AppMenu() {
 
   const menuList = [
     {
-      group: "General",
+      group: 'General',
       items: [
         {
-          id: "homebutton",
-          link: "/",
+          id: 'homebutton',
+          link: '/',
           icon: <HomeIcon />,
-          text: "Home"
-        },
-        {
-          id: "componentoverviewbutton",
-          link: "/components",
-          icon: <ReceiptText />,
-          text: "Components"
-        },
-        {
-          id: "findcomponentbutton",
-          link: "/findcomponent",
-          icon: <QrCode />,
-          text: "Find Component"
+          text: 'Home'
         },
       ]
       },
       {
-        group: "Misc",
+        group: 'Components',
         items: [
           {
-            id: "settingsbutton",
-            link: "/settings",
-            icon: <Settings />,
-            text: "Settings"
+            id: 'componentsbutton',
+            link: '/components',
+            icon: <Puzzle />,
+            text: 'Components'
           },
           {
-            id: "privacybutton",
-            link: "/",
-            icon: <FileLock />,
-            text: "Privacy"
+            id: 'findcomponentbutton',
+            link: '/findcomponent',
+            icon: <QrCode />,
+            text: 'Find Component'
+          },
+        ]
+      },
+      {
+        group: 'Models',
+        items: [
+          {
+            id: 'modelsbutton',
+            link: '/',
+            icon: <Package />,
+            text: 'Models'
+          },
+        ]
+      },
+      {
+        group: 'Misc',
+        items: [
+          // {
+          //   id: 'settingsbutton',
+          //   link: '/settings',
+          //   icon: <Settings />,
+          //   text: 'Settings'
+          // },
+          {
+            id: 'creditsbutton',
+            link: '/credits',
+            icon: <Asterisk />,
+            text: 'Credits'
           },
           {
-            id: "creditsbutton",
-            link: "/credits",
-            icon: <BookUser />,
-            text: "Credits"
+            id: 'imprintbutton',
+            link: '/',
+            icon: <BookText />,
+            text: 'Imprint'
           },
         ]
       }
@@ -57,41 +73,67 @@ export default function AppMenu() {
 
   return (
     <>
-    <div className="flex flex-col h-full justify-between">
-      <div className="flex grow">
-          <Command style={{ overflow: 'visible' }} className="rounded-lg">
+    <div className='flex flex-col h-full justify-items-center'>
+      <div className='flex grow'>
+          <Command style={{ overflow: 'visible' }} className='rounded-lg'>
             <CommandList style={{ overflow: 'visible' }}>
+
                 <CommandGroup heading={menuList[0].group}>
                   {menuList[0].items.map((option: any, optionKey: number) => 
                     <Link key={optionKey} href={option.link}>
-                      <CommandItem id={option.id} key={optionKey} className="flex gap-2 cursor-pointer">
+                      <CommandItem id={option.id} key={optionKey} className='flex gap-2 cursor-pointer'>
                         {option.icon}
                         {option.text}
                       </CommandItem>
                     </Link>
                   )}
                 </CommandGroup>
+
+                <CommandGroup heading={menuList[1].group}>
+                  {menuList[1].items.map((option: any, optionKey: number) => 
+                    <Link key={optionKey} href={option.link}>
+                      <CommandItem id={option.id} key={optionKey} className='flex gap-2 cursor-pointer'>
+                        {option.icon}
+                        {option.text}
+                      </CommandItem>
+                    </Link>
+                  )}
+                </CommandGroup>
+
+                <CommandGroup heading={menuList[2].group}>
+                  {menuList[2].items.map((option: any, optionKey: number) => 
+                    <Link key={optionKey} href={option.link}>
+                      <CommandItem id={option.id} key={optionKey} className='flex gap-2 cursor-pointer'>
+                        {option.icon}
+                        {option.text}
+                      </CommandItem>
+                    </Link>
+                  )}
+                </CommandGroup>
+
             </CommandList>
           </Command>
         </div>
 
-        <div className="flex shrink-0">
-          <Command style={{ overflow: 'visible' }} className="rounded-lg">
+        <div className='flex shrink-0'>
+          <Command style={{ overflow: 'visible' }} className='rounded-lg'>
             <CommandList style={{ overflow: 'visible' }}>
-                <CommandGroup heading={menuList[1].group}>
-                  {menuList[1].items.map((option: any, optionKey: number) => 
+
+                <CommandGroup heading={menuList[3].group}>
+                  {menuList[3].items.map((option: any, optionKey: number) => 
                     <Link key={optionKey} href={option.link}>
-                      <CommandItem id={option.id} key={optionKey} className="flex gap-2">
+                      <CommandItem id={option.id} key={optionKey} className='flex gap-2'>
                         {option.icon}
                         {option.text}
                       </CommandItem>
                     </Link>
                   )}
                 </CommandGroup>
+                
             </CommandList>
           </Command>
         </div>
       </div>
       </>
-  );
+  )
 }