+++ src/frontend/components/ComponentLookup.tsx
@@ -18,7 +18,7 @@ const ComponentLookup = () => {
     formatsToSupport: [Html5QrcodeSupportedFormats.QR_CODE]
   }
 
-    // { facingMode: "environment" }
+    // { facingMode: 'environment' }
     //  && html5QrCode != null
 
   const [referenceID, setReferenceID] = useState('')
@@ -28,12 +28,12 @@ const ComponentLookup = () => {
   const [isScanning, setIsScanning] = useState(false)
   const [borderColor, setBorderColor] = useState('silver')
   const html5QrCodeRef: MutableRefObject<Html5Qrcode | null> = useRef(null)
-  const elementId = "reader"
-  const cameraContainerId = "cameracontainer"
-  const msg_match: string = "IDs Match!"
-  const msg_mismatch: string = "Does not Match with Reference ID!"
-  const msg_ref_scanned: string = "Reference ID set.\nStart scan for comparison."
-  const msg_camera_feed: string = "Camera Feed Placeholder.\n\nTo start comparing IDs,\neither scan or type an ID as reference."
+  const elementId = 'reader'
+  const cameraContainerId = 'cameracontainer'
+  const msg_match: string = 'IDs Match!'
+  const msg_mismatch: string = 'Does not Match with Reference ID!'
+  const msg_ref_scanned: string = 'Reference ID set.\nStart scan for comparison.'
+  const msg_camera_feed: string = 'Camera Feed Placeholder.\n\nTo start comparing IDs,\neither scan or type an ID as reference.'
 
   const startScanningForReference = () => {
       document.getElementById(elementId)?.scrollIntoView()
@@ -42,11 +42,11 @@ const ComponentLookup = () => {
       }
       if (html5QrCodeRef.current && !referenceID && !isScanning) {
           setIsScanning(true)
-          console.log("Starting scan for reference QR code.")
+          console.log('Starting scan for reference QR code.')
           Html5Qrcode.getCameras().then(cameras => {
               if (cameras.length && html5QrCodeRef.current != null) {
                 html5QrCodeRef.current.start(
-                    { facingMode: "environment" },
+                    { facingMode: 'environment' },
                       config,
                       decodedText => {
                           setReferenceID(decodedText)
@@ -56,11 +56,11 @@ const ComponentLookup = () => {
                       },
                       undefined
                       // onScanFailure
-                  ).catch(err => console.error("Error starting QR scan: ", err))
+                  ).catch(err => console.error('Error starting QR scan: ', err))
               } else {
-                  console.error("No cameras found.")
+                  console.error('No cameras found.')
               }
-          }).catch(err => console.error("Error getting cameras: ", err))
+          }).catch(err => console.error('Error getting cameras: ', err))
       }
   }
 
@@ -71,11 +71,11 @@ const ComponentLookup = () => {
       }
       if (html5QrCodeRef.current && referenceID && !isScanning) {
           setIsScanning(true)
-          console.log("Starting scan for comparison QR code.")
+          console.log('Starting scan for comparison QR code.')
           Html5Qrcode.getCameras().then(cameras => {
               if (cameras.length && html5QrCodeRef.current != null) {
                 html5QrCodeRef.current.start(
-                    { facingMode: "environment" },
+                    { facingMode: 'environment' },
                       config,
                       decodedText => {
                           const resultMessage: string = decodedText === referenceID ? msg_match : msg_mismatch
@@ -92,11 +92,11 @@ const ComponentLookup = () => {
                       },
                       undefined
                       // onScanFailure
-                  ).catch(err => console.error("Error starting QR scan: ", err))
+                  ).catch(err => console.error('Error starting QR scan: ', err))
               } else {
-                  console.error("No cameras found.")
+                  console.error('No cameras found.')
               }
-          }).catch(err => console.error("Error getting cameras: ", err))
+          }).catch(err => console.error('Error getting cameras: ', err))
       }
   }
 
@@ -105,9 +105,9 @@ const ComponentLookup = () => {
           setIsScanning(false)
           // setBorderColor('silver')
           html5QrCodeRef.current?.clear()
-          console.log("Scanning stopped.")
+          console.log('Scanning stopped.')
       }).catch(err => {
-          console.error("Failed to stop scanning.", err)
+          console.error('Failed to stop scanning.', err)
       })
 }
 
@@ -133,20 +133,20 @@ const ComponentLookup = () => {
   }
 
   return (
-    <div className="flex flex-col items-center">
-      <CardHeader className="p-1 relative w-full items-center flex max-w-sm">
+    <div className='flex flex-col items-center'>
+      <CardHeader className='p-1 relative w-full items-center flex max-w-sm'>
           {/* Set Reference ID Interface */}
           {!referenceID && !isScanning &&
-            <div className="flex w-full max-w-sm items-center space-x-2 pb-4">
+            <div className='flex w-full max-w-sm items-center space-x-2 pb-4'>
               <Input
-                id="inputFieldReferenceID"
-                placeholder="Reference ID"
+                id='inputFieldReferenceID'
+                placeholder='Reference ID'
                 value={inputReferenceID}
                 onChange={e => setInputReferenceID(e.target.value)}
               />
               <Button
-                variant="outline"
-                className="w-[200px] hover:bg-[#009cda] hover:text-white"
+                variant='outline'
+                className='w-[200px] hover:bg-[#009cda] hover:text-white'
                 onClick={handleSetInputReferenceID}>
                   Set Ref. ID
               </Button>
@@ -154,26 +154,26 @@ const ComponentLookup = () => {
           }
 
         {/* Display Reference and Current ID */}
-        <div className="w-full grid grid-cols-2 gap-y-4 items-center">
+        <div className='w-full grid grid-cols-2 gap-y-4 items-center'>
           {/* Row 1 */}
-          <div className="flex justify-start min-w-0 w-20">
-            <p className="text-lg">Ref. ID: </p>
+          <div className='flex justify-start min-w-0 w-20'>
+            <p className='text-lg'>Ref. ID: </p>
           </div>
-          <div className="flex justify-end items-center">
+          <div className='flex justify-end items-center'>
             <Badge
-              variant="secondary"
-              className={referenceID? "bg-[#009cda] hover:bg-[#009cda] text-white no-wrap flex-shrink-0" : "bg-[#c0c0c0]	hover:bg-[#c0c0c0] no-wrap flex-shrink-0"}>
+              variant='secondary'
+              className={referenceID? 'bg-[#009cda] hover:bg-[#009cda] text-white no-wrap flex-shrink-0' : 'bg-[#c0c0c0]	hover:bg-[#c0c0c0] no-wrap flex-shrink-0'}>
                 {referenceID || 'Not set'}
             </Badge>
           </div>
           {/* Row 2 */}
-          <div className="flex justify-start min-w-0 w-20">
-            <p className="text-lg">Cur. ID: </p>
+          <div className='flex justify-start min-w-0 w-20'>
+            <p className='text-lg'>Cur. ID: </p>
           </div>
-          <div className="flex justify-end items-center">
+          <div className='flex justify-end items-center'>
             <Badge
-              variant="secondary"
-              className="no-wrap flex-shrink-0"
+              variant='secondary'
+              className='no-wrap flex-shrink-0'
               style={{backgroundColor: `${borderColor}`}}>
                 {currentID || 'Not set'}
             </Badge>
@@ -184,26 +184,26 @@ const ComponentLookup = () => {
       {/* QR Code Scanner Canvas */}
       <CardContent
         id={cameraContainerId}
-        className="mt-4 p-0 relative w-full max-w-[500px] h-[500px] border-8 rounded-xl"
+        className='mt-4 p-0 relative w-full max-w-[500px] h-[500px] border-8 rounded-xl'
         style={{borderColor: `${borderColor}`}}>
           {/* QR Code Scanner Element */}
-          <div id={elementId} className="rounded-lg"></div>
+          <div id={elementId} className='rounded-lg'></div>
           {!isScanning &&
-            <div className="absolute inset-0 bg-gray-200 flex items-center text-center justify-center rounded">
-              <span className="whitespace-pre-wrap">
+            <div className='absolute inset-0 bg-gray-200 flex items-center text-center justify-center rounded'>
+              <span className='whitespace-pre-wrap'>
                 {comparisonResult? comparisonResult : msg_camera_feed}
               </span>
             </div>
           }
       </CardContent>
 
-      <div className="m-4 flex space-y-4 flex-col items-center">
+      <div className='m-4 flex space-y-4 flex-col items-center'>
         {/* Start Reference Scan Button */}
         {!referenceID &&
           <Button
             onClick={startScanningForReference}
-            variant="outline"
-            className="w-[200px] hover:bg-[#009cda] hover:text-white">
+            variant='outline'
+            className='w-[200px] hover:bg-[#009cda] hover:text-white'>
               Start QR Code Scan
           </Button>
         }
@@ -211,8 +211,8 @@ const ComponentLookup = () => {
         {referenceID && !isScanning &&
           <Button
             onClick={startScanningForComparison}
-            variant="outline"
-            className="w-[200px] hover:bg-[#009cda] hover:text-white">
+            variant='outline'
+            className='w-[200px] hover:bg-[#009cda] hover:text-white'>
               Start QR Code Scan
           </Button>
         }
@@ -220,8 +220,8 @@ const ComponentLookup = () => {
         {isScanning &&
           <Button
             onClick={stopScanning}
-            variant="outline"
-            className="w-[200px] hover:bg-[#009cda] hover:text-white">
+            variant='outline'
+            className='w-[200px] hover:bg-[#009cda] hover:text-white'>
               Stop Scanning
           </Button>
         }
@@ -229,8 +229,8 @@ const ComponentLookup = () => {
         {!isScanning &&
           <Button
             onClick={resetScanner}
-            variant="outline"
-            className="m-4 w-[200px] bg-red-500 hover:bg-red-600 hover:text-white">
+            variant='outline'
+            className='m-4 w-[200px] bg-red-500 hover:bg-red-600 hover:text-white'>
               Reset
           </Button>
         }