+++ src/frontend/components/Header.tsx
@@ -1,18 +1,18 @@
-'use client';
+'use client'
 
-import { BookOpenText, Menu } from "lucide-react";
-import AppMenu from "./AppMenu";
+import { BookOpenText, Menu } from 'lucide-react'
+import AppMenu from './AppMenu'
 
 export default function Header() {
   return (
     <>
-      <div className="flex gap-4 p-4 border-b font-bold grow">
+      <div className='flex gap-4 p-4 border-b font-bold grow'>
         <BookOpenText /><h1>Catalogue of Second Chances</h1>
       </div>
 
-      <div className="md:hidden p-4">
+      <div className='md:hidden p-4'>
         <AppMenu />
       </div>
     </>
-  );
+  )
 }