+++ src/frontend/components/Sidebar.tsx
@@ -1,12 +1,12 @@
 'use client'
 
-import UserItem from "./UserItem";
-import AppMenu from "./AppMenu";
+import UserItem from './UserItem'
+import AppMenu from './AppMenu'
 
 export default function Sidebar() {
 
   return (
-    <div className="fixed flex flex-col gap-4 w-[300px] min-w-[300px] p-4 h-full justify-between">
+    <div className='fixed flex flex-col gap-4 w-[250px] min-w-[250px] p-4 h-[95dvh] justify-between'>
       
       {/* <div>
         <UserItem/>
@@ -15,5 +15,5 @@ export default function Sidebar() {
       <AppMenu />
 
     </div>
-  );
+  )
 }