+++ src/frontend/components/UserItem.tsx
@@ -1,12 +1,13 @@
-'use client';
+'use client'
+
 export default function UserItem() {
-    return <div className="flex items-center justify-between gap-2 border rounded-[8px] p-2">
-        <div className="avatar rounded-full min-h-10 min-w-10 bg-[#009cda] text-white font-[700] flex items-center justify-center">
+    return <div className='flex items-center justify-between gap-2 border rounded-[8px] p-2'>
+        <div className='avatar rounded-full min-h-10 min-w-10 bg-[#009cda] text-white font-[700] flex items-center justify-center'>
             <p>MES</p>
         </div>
-        <div className="grow">
-            <p className="text-[16px] font-bold">Max Eschenbach</p>
-            <p className="text-[12px] text-neutral-500">eschenbach@dg.tu-darmstadt.de</p>
+        <div className='grow'>
+            <p className='text-[16px] font-bold'>Max Eschenbach</p>
+            <p className='text-[12px] text-neutral-500'>eschenbach@dg.tu-darmstadt.de</p>
         </div>
-    </div>;
+    </div>
 }